+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Added
 - Add method to get seller by id  in catalog client
+## [2.6.0] - 2021-02-11
+### Added
+- Adding RatesAndBenefits Client (https://developers.vtex.com/vtex-developer-docs/reference/benefits)
 
 ## [2.5.3-beta] - 2021-02-08
 ### Fixed